+++ packages/brownie/release.config.cjs
@@ -1,4 +1,6 @@
+// publish
 module.exports = {
+  ...require('../../release.config.cjs'),
   plugins: [
     '@semantic-release/commit-analyzer',
     '@semantic-release/release-notes-generator',
@@ -9,13 +11,7 @@ module.exports = {
         assets: ['CHANGELOG.md', 'README.md', 'docs/']
       }
     ],
-    [
-      '@semantic-release/gitlab',
-      {
-        gitlabUrl: 'https://gitlab.tailored-apps.com',
-        assets: []
-      }
-    ],
+    '@semantic-release/gitlab',
     '@semantic-release/npm'
   ]
 }