+++ packages/ts-utility-types/release.config.cjs
@@ -1,4 +1,6 @@
+// publish
 module.exports = {
+  ...require('../../release.config.cjs'),
   plugins: [
     '@semantic-release/commit-analyzer',
     '@semantic-release/release-notes-generator',
@@ -6,16 +8,10 @@ module.exports = {
     [
       '@semantic-release/git',
       {
-        assets: [ 'CHANGELOG.md', 'README.md', 'docs/' ]
+        assets: ['CHANGELOG.md', 'README.md', 'docs/']
       }
     ],
-    [
-      '@semantic-release/gitlab',
-      {
-        gitlabUrl: 'https://gitlab.tailored-apps.com',
-        assets: []
-      }
-    ],
+    '@semantic-release/gitlab',
     '@semantic-release/npm'
   ]
 }