+++ packages/nestjs-graphql-typeorm-dataloader/CHANGELOG.md
@@ -1,3 +1,15 @@
+# @webundsoehne/nestjs-graphql-typeorm-dataloader [3.3.0-beta.1](https://gitlab.tailored-apps.com/bdsm/nx-skeleton/compare/@webundsoehne/nestjs-graphql-typeorm-dataloader@3.2.3...@webundsoehne/nestjs-graphql-typeorm-dataloader@3.3.0-beta.1) (2022-04-19)
+
+
+### Bug Fixes
+
+* update base package to esm for new multi-semantic-release versions ([d0798b8](https://gitlab.tailored-apps.com/bdsm/nx-skeleton/commit/d0798b822e3f25968a573712b998a189292159c7))
+
+
+### Features
+
+* get over semantic-release hang ([5f83281](https://gitlab.tailored-apps.com/bdsm/nx-skeleton/commit/5f83281a01242f536ce2b57a98f049c42c6684ab))
+
 ## @webundsoehne/nestjs-graphql-typeorm-dataloader [3.2.3](https://gitlab.tailored-apps.com/bdsm/nx-skeleton/compare/@webundsoehne/nestjs-graphql-typeorm-dataloader@3.2.2...@webundsoehne/nestjs-graphql-typeorm-dataloader@3.2.3) (2022-04-07)
 
 ### Bug Fixes