+++ packages/nestjs-keycloak/CHANGELOG.md
@@ -1,3 +1,21 @@
+# @webundsoehne/nestjs-keycloak [2.3.0-beta.1](https://gitlab.tailored-apps.com/bdsm/nx-skeleton/compare/@webundsoehne/nestjs-keycloak@2.2.0...@webundsoehne/nestjs-keycloak@2.3.0-beta.1) (2022-04-19)
+
+
+### Bug Fixes
+
+* bump release ([ab310c3](https://gitlab.tailored-apps.com/bdsm/nx-skeleton/commit/ab310c3c2f0db48bdbf9fff9c31ff4f171055e01))
+* fixes previous publishing fix ([4365bbe](https://gitlab.tailored-apps.com/bdsm/nx-skeleton/commit/4365bbe69e840e563097dd629c1cd9352a89956c))
+* housekeeping, swap links to oss repos on github ([7e0f209](https://gitlab.tailored-apps.com/bdsm/nx-skeleton/commit/7e0f2093cf8c9afddd3d2f9228ec81896eb2d5e6))
+* janitoring, README.md updates ([fe8a4c4](https://gitlab.tailored-apps.com/bdsm/nx-skeleton/commit/fe8a4c47bba245fe6338988fef4c5c025e455666))
+* linters gonna lint ([6d7e3e0](https://gitlab.tailored-apps.com/bdsm/nx-skeleton/commit/6d7e3e0bfee76e363ab9868d1244adab5a3546ec))
+* republish packages ([3f2b2e3](https://gitlab.tailored-apps.com/bdsm/nx-skeleton/commit/3f2b2e339fc78c4e11263981fc78d787034472ff))
+* update base package to esm for new multi-semantic-release versions ([d0798b8](https://gitlab.tailored-apps.com/bdsm/nx-skeleton/commit/d0798b822e3f25968a573712b998a189292159c7))
+
+
+### Features
+
+* get over semantic-release hang ([5f83281](https://gitlab.tailored-apps.com/bdsm/nx-skeleton/commit/5f83281a01242f536ce2b57a98f049c42c6684ab))
+
 # @webundsoehne/nestjs-keycloak [2.2.0](https://github.com/tailoredmedia/backend-nx-skeleton/compare/@webundsoehne/nestjs-keycloak@2.1.0...@webundsoehne/nestjs-keycloak@2.2.0) (2022-02-17)
 
 ### Features